fix: sync openapi spec (#1403)
* fix: sync openapi spec

* feat: add labels to multiple create and delete methods

* feat: print labels for executors and webhooks

* fix: add method to run multiple executions

* feat: cli to run multiple test and test suites

* feat: generics based worker pool

* feat: unit tests for worker pool

* fix: go fmt

* feat: api to run multiple tests and test suites

* fix: change cli run command printing conditions

* feat: use concurrency parameter

* feat: delete tests and test suites byt labels

* fix: support sync test execution

* Update internal/app/api/v1/webhook.go


* fix: helm golint

* fix: error golint

* fix: golint server call

* fix: govet for mongodb

* fix: return labels for webhooks and executors

* fix: webhook delete

* fix: tune results reading

* fix: propagate labels for multiple test and test suite executions

* fix: use proper test name

* fix: goling

* fix: golint

* fix: add missing param for test update

* feat: add test suite executions labels to list view

* fix: use background context

* feat: doc update

* fix: dep update

+++ cmd/kubectl-testkube/commands/executors/delete.go
@@ -1,33 +1,43 @@
 package executors
 
 import (
+	"strings"
+
 	"github.com/kubeshop/testkube/cmd/kubectl-testkube/commands/common"
-	"github.com/kubeshop/testkube/cmd/kubectl-testkube/commands/common/validator"
 	"github.com/kubeshop/testkube/pkg/ui"
 	"github.com/spf13/cobra"
 )
 
 func NewDeleteExecutorCmd() *cobra.Command {
 	var name string
+	var selectors []string
 
 	cmd := &cobra.Command{
 		Use:   "executor [executorName]",
 		Short: "Delete Executor",
 		Long:  `Delete Executor Resource, pass name to delete by name`,
-		Args:  validator.ExecutorName,
 		Run: func(cmd *cobra.Command, args []string) {
-			name = args[0]
-
 			client, _ := common.GetClient(cmd)
+			if len(args) > 0 {
+				name = args[0]
+				err := client.DeleteExecutor(name)
+				ui.ExitOnError("deleting executor: "+name, err)
+				ui.SuccessAndExit("Succesfully deleted executor", name)
+			}
 
-			err := client.DeleteExecutor(name)
-			ui.ExitOnError("deleting executor: "+name, err)
+			if len(selectors) != 0 {
+				selector := strings.Join(selectors, ",")
+				err := client.DeleteExecutors(selector)
+				ui.ExitOnError("deleting executors by labels: "+selector, err)
+				ui.SuccessAndExit("Succesfully deleted executors by labels", selector)
+			}
 
-			ui.Success("Executor deleted")
+			ui.Failf("Pass Executor name or labels to delete by labels")
 		},
 	}
 
 	cmd.Flags().StringVarP(&name, "name", "n", "", "unique executor name, you can also pass it as first argument")
+	cmd.Flags().StringSliceVarP(&selectors, "label", "l", nil, "label key value pair: --label key1=value1")
 
 	return cmd
 }
